requirements.txt
flask
pytest
psycopg2-binary
flask-sqlalchemy
python-dotenv
pytest-cov